# public/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ServicePro API",
    "description": "REST API for the ServicePro service business management platform. Manage customers, jobs, estimates, invoices, and contracts programmatically.",
    "version": "1.0.0",
    "contact": {
      "name": "ServicePro Support"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "ServicePro API v1 (relative)"
    }
  ],
  "security": [
    { "BearerAuth": [] },
    { "ApiKeyAuth": [] }
  ],
...
  "paths": {
...
    "/estimates": {
      "get": {
        "tags": ["Estimates"],
        "operationId": "listEstimates",
        "summary": "List estimates",
        "description": "Returns a paginated list of estimates belonging to the authenticated user's company. Supports search, filtering by status or customer, and sorting.",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Search" },
          {
            "name": "status",
            "in": "query",
            "description": "Filter estimates by status.",
            "schema": { "type": "string", "enum": ["draft", "sent", "viewed", "approved", "rejected", "expired"] }
          },
          {
            "name": "customerId",
            "in": "query",
            "description": "Filter estimates by the associated customer ID.",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of estimates.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/PaginatedResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Estimate" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["Estimates"],
        "operationId": "createEstimate",
        "summary": "Create an estimate",
        "description": "Creates a new estimate for a customer. The customerId field is required. An estimate number is automatically generated.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EstimateCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created estimate.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Estimate" }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/invoices/{id}": {
      "get": {
        "tags": ["Invoices"],
        "operationId": "getInvoice",
        "summary": "Get an invoice",
        "description": "Returns a single invoice by ID, including all line items, calculated totals, and payment history. The invoice must belong to the authenticated user's company.",
        "parameters": [
          { "$ref": "#/components/parameters/ResourceId" }
        ],
        "responses": {
          "200": {
            "description": "The requested invoice.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Invoice" }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "patch": {
        "tags": ["Invoices"],
        "operationId": "updateInvoice",
        "summary": "Update an invoice",
        "description": "Partially updates an existing invoice. Only the fields provided in the request body will be changed. The invoice must belong to the authenticated user's company.",
        "parameters": [
          { "$ref": "#/components/parameters/ResourceId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InvoiceUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated invoice.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Invoice" }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "tags": ["Invoices"],
        "operationId": "deleteInvoice",
        "summary": "Delete an invoice",
        "description": "Permanently deletes an invoice record. The invoice must belong to the authenticated user's company. This action cannot be undone.",
        "parameters": [
          { "$ref": "#/components/parameters/ResourceId" }
        ],
        "responses": {
          "200": {
            "description": "Confirmation that the invoice was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid", "description": "The ID of the deleted invoice." },
                        "deleted": { "type": "boolean", "description": "Always true on successful deletion.", "example": true }
                      },
                      "required": ["id", "deleted"]
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Authenticate using a Bearer token in the Authorization header. Tokens start with `sb_live_`."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Authenticate using an API key passed in the X-API-Key header. Keys start with `sb_live_`."
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "description": "Page number for pagination (1-based).",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "description": "Number of items per page. Maximum is 100.",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
...
      "Search": {
        "name": "q",
        "in": "query",
        "description": "Search query string. Searches across relevant text fields for the resource.",
        "schema": { "type": "string" }
      },
      "ResourceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "The unique identifier (UUID) of the resource.",
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "schemas": {
      "Pagination": {
        "type": "object",
        "description": "Pagination metadata included with list responses.",
        "properties": {
          "page": { "type": "integer", "description": "The current page number.", "example": 1 },
          "limit": { "type": "integer", "description": "The number of items per page.", "example": 20 },
          "total": { "type": "integer", "description": "The total number of items across all pages.", "example": 57 }
        },
        "required": ["page", "limit", "total"]
      },
      "PaginatedResponse": {
        "type": "object",
        "description": "Wrapper for paginated list responses.",
        "properties": {
          "data": {
            "type": "array",
            "description": "The array of resource items for this page.",
            "items": {}
          },
          "pagination": { "$ref": "#/components/schemas/Pagination" }
        },
        "required": ["data", "pagination"]
      },
      "Error": {
        "type": "object",
        "description": "Standard error response returned when a request fails.",
        "properties": {
          "error": { "type": "string", "description": "A human-readable error message describing what went wrong.", "example": "Not authenticated" }
        },
        "required": ["error"]
      },
...
      "Estimate": {
        "type": "object",
        "description": "An estimate record with line items, pricing, and terms sent to a customer for approval.",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier for the estimate." },
          "customerId": { "type": "string", "format": "uuid", "description": "ID of the customer this estimate is for." },
          "customerName": { "type": "string", "description": "Name of the associated customer (read-only).", "nullable": true, "example": "Jane Smith" },
          "estimateNumber": { "type": "string", "description": "Auto-generated estimate reference number.", "example": "EST-0042" },
          "status": { "type": "string", "enum": ["draft", "sent", "viewed", "approved", "rejected", "expired"], "description": "Current status of the estimate.", "example": "sent" },
          "issueDate": { "type": "string", "format": "date", "description": "Date the estimate was issued.", "nullable": true, "example": "2026-02-06" },
          "expiryDate": { "type": "string", "format": "date", "description": "Date the estimate expires.", "nullable": true, "example": "2026-03-06" },
          "lineItems": { "type": "array", "description": "List of products or services included in the estimate.", "items": { "$ref": "#/components/schemas/LineItem" } },
          "subtotal": { "type": "number", "description": "Sum of all line item amounts before tax.", "example": 450.00 },
          "taxRate": { "type": "number", "description": "Tax rate as a percentage (e.g. 8.5 for 8.5%).", "nullable": true, "example": 8.5 },
          "taxAmount": { "type": "number", "description": "Calculated tax amount in dollars.", "example": 38.25 },
          "total": { "type": "number", "description": "Grand total including tax.", "example": 488.25 },
          "notes": { "type": "string", "description": "Notes visible to the customer on the estimate.", "nullable": true },
          "terms": { "type": "string", "description": "Terms and conditions for the estimate.", "nullable": true },
          "address": { "type": "string", "description": "Service address for the estimate.", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time", "description": "Timestamp when the estimate was created." },
          "updatedAt": { "type": "string", "format": "date-time", "description": "Timestamp when the estimate was last updated." }
        },
        "required": ["id", "customerId", "estimateNumber", "status", "lineItems", "subtotal", "total", "createdAt", "updatedAt"]
      },
      "EstimateCreate": {
        "type": "object",
        "description": "Input for creating a new estimate. The customerId is required. An estimate number is auto-generated.",
        "properties": {
          "customerId": { "type": "string", "format": "uuid", "description": "ID of the customer this estimate is for." },
          "lineItems": { "type": "array", "description": "List of line items for the estimate.", "items": { "$ref": "#/components/schemas/LineItemInput" } },
          "taxRate": { "type": "number", "description": "Tax rate as a percentage (e.g. 8.5 for 8.5%).", "example": 8.5 },
          "issueDate": { "type": "string", "format": "date", "description": "Date the estimate is issued. Defaults to today.", "example": "2026-02-06" },
          "expiryDate": { "type": "string", "format": "date", "description": "Date the estimate expires.", "example": "2026-03-06" },
          "notes": { "type": "string", "description": "Notes visible to the customer." },
          "terms": { "type": "string", "description": "Terms and conditions." },
          "address": { "type": "string", "description": "Service address for the estimate." },
          "status": { "type": "string", "enum": ["draft", "sent", "viewed", "approved", "rejected", "expired"], "description": "Initial estimate status. Defaults to 'draft'.", "default": "draft" }
        },
        "required": ["customerId"]
      },
...
      "Invoice": {
        "type": "object",
        "description": "An invoice record with line items, pricing, payment tracking, and terms.",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier for the invoice." },
          "customerId": { "type": "string", "format": "uuid", "description": "ID of the customer this invoice is for." },
          "customerName": { "type": "string", "description": "Name of the associated customer (read-only).", "nullable": true, "example": "Jane Smith" },
          "jobId": { "type": "string", "format": "uuid", "description": "ID of the job this invoice is linked to.", "nullable": true },
          "estimateId": { "type": "string", "format": "uuid", "description": "ID of the estimate this invoice was created from.", "nullable": true },
          "invoiceNumber": { "type": "string", "description": "Auto-generated invoice reference number.", "example": "INV-0078" },
          "status": { "type": "string", "enum": ["draft", "sent", "viewed", "paid", "partial", "overdue", "cancelled"], "description": "Current status of the invoice.", "example": "sent" },
          "issueDate": { "type": "string", "format": "date", "description": "Date the invoice was issued.", "nullable": true, "example": "2026-02-06" },
          "dueDate": { "type": "string", "format": "date", "description": "Date the invoice payment is due.", "nullable": true, "example": "2026-03-06" },
          "lineItems": { "type": "array", "description": "List of products or services billed on the invoice.", "items": { "$ref": "#/components/schemas/LineItem" } },
          "subtotal": { "type": "number", "description": "Sum of all line item amounts before tax.", "example": 450.00 },
          "taxRate": { "type": "number", "description": "Tax rate as a percentage.", "nullable": true, "example": 8.5 },
          "taxAmount": { "type": "number", "description": "Calculated tax amount in dollars.", "example": 38.25 },
          "total": { "type": "number", "description": "Grand total including tax.", "example": 488.25 },
          "amountPaid": { "type": "number", "description": "Total amount paid so far.", "example": 200.00 },
          "amountDue": { "type": "number", "description": "Remaining balance (total minus amountPaid).", "example": 288.25 },
          "payments": { "type": "array", "description": "List of payment records for this invoice.", "items": { "$ref": "#/components/schemas/Payment" }, "nullable": true },
          "notes": { "type": "string", "description": "Notes visible to the customer on the invoice.", "nullable": true },
          "terms": { "type": "string", "description": "Payment terms and conditions.", "nullable": true },
          "address": { "type": "string", "description": "Billing or service address.", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time", "description": "Timestamp when the invoice was created." },
          "updatedAt": { "type": "string", "format": "date-time", "description": "Timestamp when the invoice was last updated." }
        },
        "required": ["id", "customerId", "invoiceNumber", "status", "lineItems", "subtotal", "total", "amountPaid", "amountDue", "createdAt", "updatedAt"]
      },
...
      "InvoiceUpdate": {
        "type": "object",
        "description": "Input for partially updating an existing invoice. All fields are optional.",
        "properties": {
          "jobId": { "type": "string", "format": "uuid", "description": "ID of the job to link this invoice to." },
          "estimateId": { "type": "string", "format": "uuid", "description": "ID of the estimate this invoice was created from." },
          "lineItems": { "type": "array", "description": "Replacement list of line items (replaces all existing).", "items": { "$ref": "#/components/schemas/LineItemInput" } },
          "taxRate": { "type": "number", "description": "Tax rate as a percentage." },
          "issueDate": { "type": "string", "format": "date", "description": "Date the invoice is issued." },
          "dueDate": { "type": "string", "format": "date", "description": "Date the invoice payment is due." },
          "notes": { "type": "string", "description": "Notes visible to the customer." },
          "terms": { "type": "string", "description": "Payment terms and conditions." },
          "address": { "type": "string", "description": "Billing or service address." },
          "status": { "type": "string", "enum": ["draft", "sent", "viewed", "paid", "partial", "overdue", "cancelled"], "description": "Updated invoice status." }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request was invalid. Check the error message for details on what needs to be corrected.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Title must be at least 2 characters" }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication failed. Provide a valid Bearer token or X-API-Key header.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Not authenticated" }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource was not found or does not belong to your company.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Resource not found" }
          }
        }
      },
      "InternalError": {
        "description": "An unexpected server error occurred. If the problem persists, contact support.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Internal server error" }
          }
        }
      }
    }
  }
}